+++ R/HermesData-class.R
@@ -106,6 +106,16 @@ HermesData <- function(object) { # nolint
     not_empty(assays(object))
   )
 
+  assays(object) <- lapply(assays(object), function(x) {
+    if ("DelayedMatrix" %in% class(x)) {
+      x <- as.matrix(x)
+      mode(x) <- "integer"
+      x
+    } else {
+      x
+    }
+  })
+
   missing_row <- setdiff(.row_data_cols, names(rowData(object)))
   rowData(object)[, missing_row] <- NA