+++ tests/testthat/test-HermesData-class.R
@@ -91,6 +91,26 @@ test_that("HermesData creates missing columns with NAs correctly", {
   expect_true(all(.row_data_cols %in% names(rowData(result))))
 })
 
+test_that("HermesData converts DelayedMatrix assays to matrix/assay type", {
+  n_row = length(rownames(SummarizedExperiment::rowData(hermes_data)))
+  n_col = length(colnames(SummarizedExperiment::assay(hermes_data)))
+  vals <- rnorm(n = n_row * n_col, mean = 10, sd = 20)
+  vals_m <- round(matrix(vals, ncol = n_col))
+  vals_m <- abs(vals_m)
+  rownames(vals_m) <- rownames(SummarizedExperiment::rowData(hermes_data))
+  colnames(vals_m) <- rownames(SummarizedExperiment::colData(hermes_data))
+  mock_delayed_array <- DelayedArray::DelayedArray(vals_m)
+  my_se <- SummarizedExperiment(
+    rowData = SummarizedExperiment::rowData(hermes_data),
+    colData = SummarizedExperiment::colData(hermes_data)
+  )
+  SummarizedExperiment::assays(my_se) <- SimpleList(mock_delayed_array, mock_delayed_array)
+  SummarizedExperiment::assayNames(my_se) <- c("counts", "not_counts")
+  my_hd <- HermesData(my_se)
+
+  expect_true(all(unlist(lapply(assays(my_hd), function(x) c("matrix", "array") %in% class(x)))))
+})
+
 test_that("RangedHermesData objects can be created with constructor HermesData", {
   result <- expect_silent(HermesData(get_rse()))
   expect_is(result, "RangedHermesData")